+++ __tests__/sdk/best-route.ava.ts
@@ -4,7 +4,6 @@ import Big from 'big.js'
 import { MainnetRpc } from 'near-workspaces'
 import { Arbitoor, getRoutePath } from '../../sdk'
 import { InMemoryProvider } from '../../sdk/AccountProvider'
-import { getSpinMarkets } from '../../sdk/spin'
 
 test('best route', async t => {
   const user = 'test.near'
@@ -16,22 +15,17 @@ test('best route', async t => {
     return map
   }, new Map<string, TokenInfo>())
 
-  // Filter out NEAR based markets until a wrapping solution is found
-  const spinMarkets = (await getSpinMarkets(MainnetRpc))
-    .filter(market => market.base.symbol !== 'NEAR' && market.quote.symbol !== 'NEAR')
-
-  const inMemoryProvider = new InMemoryProvider(MainnetRpc, tokenMap, spinMarkets)
+  const inMemoryProvider = new InMemoryProvider(MainnetRpc, tokenMap)
 
   const arbitoor = new Arbitoor({
     accountProvider: inMemoryProvider,
-    user,
-    routeCacheDuration: 1000
+    user
   })
 
   // USDT->USN is being routed as USDT->USDC->USN on Ref, giving worse rate
-  const inputToken = 'usn'
-  const outputToken = 'a0b86991c6218b36c1d19d4a2e9eb0ce3606eb48.factory.bridge.near'
-  const inputAmount = new Big(10).pow(tokenMap.get(inputToken)!.decimals).mul(100).toString()
+  const inputToken = 'a0b86991c6218b36c1d19d4a2e9eb0ce3606eb48.factory.bridge.near'
+  const outputToken = 'usn'
+  const inputAmount = new Big(10).pow(tokenMap.get(inputToken)!.decimals).mul(1).toString()
 
   const slippageTolerance = 5
 
@@ -48,13 +42,13 @@ test('best route', async t => {
 
   for (const route of routes) {
     console.log('dex', route.dex, 'output', route.output.toString())
-    const txs = await arbitoor.generateTransactions({
-      routeInfo: route,
-      slippageTolerance
-    })
-    console.log('txs', JSON.stringify(txs, undefined, 4))
+    // const txs = await arbitoor.generateTransactions({
+    //   routeInfo: route,
+    //   slippageTolerance
+    // })
+    // console.log('txs', JSON.stringify(txs, undefined, 4))
 
-    const path = getRoutePath(route)
-    console.log('path', JSON.stringify(path, undefined, 4))
+    // const path = getRoutePath(route)
+    // console.log('path', JSON.stringify(path, undefined, 4))
   }
 })