+++ sdk/ref-finance/swap-service.ts
@@ -1,6 +1,7 @@
 import { TokenInfo } from '@tonic-foundation/token-list'
 import Big from 'big.js'
 import { SpinRouteInfo } from '../spin'
+import { TonicRouteInfo } from '../tonic'
 
 export interface ReservesMap {
   [index: string]: Big;
@@ -79,7 +80,7 @@ export interface RefRouteInfo {
 }
 
 // A route to reach token 1 to token 2
-export type RouteInfo = RefRouteInfo | SpinRouteInfo
+export type RouteInfo = RefRouteInfo | SpinRouteInfo | TonicRouteInfo
 
 export interface RefForkSwap {
   exchange: string,