+++ CHANGELOG.md
@@ -63,3 +63,27 @@
 
 - Fetch more records per orderbook.
 - Optimize orderbook caching code.
+
+## 1.2.2
+
+### Fix
+
+- Limit price for Spin swaps is now a multiple of tick size. Fixes order price precision issue.
+- Enforce quantity limits for input and output amounts on Spin.
+
+## 1.2.3
+
+### Fix
+
+- Subtract taker fee from Spin output.
+
+## 1.2.4
+
+### Breaking
+
+- Remove `routeCacheDuration` field from Arbitoor constructor.
+- Remove `spinMarkets` field from inMemoryProvider constructor.
+
+### Features
+
+- `referrer` address can be passed as a parameter.